Done heatmaps. Rearranged files.
