+++ Gradient_descent/testing.py
